Changed migrations of hunting spot, items and added world map migration, created world map syncronizer, creature dropd syncronize, created the show hunting spot page, updated hunting spot create page.

+++ backend/database/migrations/2017_11_29_053058_create_hunting_spots_table.php
@@ -18,8 +18,8 @@ class CreateHuntingSpotsTable extends Migration
             $table->string('title');
             $table->longText('description');
             $table->longText('tips')->nullable();
-            $table->decimal('experience', 10, 2);
-            $table->decimal('profit', 10, 2)->nullable();
+            $table->integer('experience');
+            $table->integer('profit')->nullable();
             $table->integer('level_min');
             $table->integer('level_max');
             $table->boolean('require_premium')->default(true);
